workspace/projects/prj001/easy_agent_builder/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "easy-agent-builder"
version = "0.1.0"
description = "Framework rápido para criação e deployment de agentes AI em GCP usando ADK"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "AI Solutions Architect", email = "architect@ai-solutions.com"}
]
keywords = ["agent", "ai", "gcp", "vertex-ai", "adk", "orchestration"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
]

dependencies = [
    # Core - ADK
    "google-adk>=0.5.0",
    
    # GCP Integration
    "google-cloud-aiplatform>=1.74.0",
    "google-cloud-storage>=2.18.0",
    "google-cloud-bigquery>=3.26.0",
    "google-cloud-logging>=3.11.0",
    "google-cloud-trace>=1.14.0",
    
    # Web Framework (for API exposure)
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    
    # CLI & Utils
    "click>=8.1.0",
    "rich>=13.9.0",
    "pydantic>=2.9.0",
    "pydantic-settings>=2.6.0",
    "pyyaml>=6.0.2",
    "jinja2>=3.1.4",
    
    # Async & Concurrency
    "aiohttp>=3.11.0",
    "tenacity>=9.0.0",
    
    # Integration
    "redis>=5.2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=6.0.0",
    "black>=24.10.0",
    "ruff>=0.7.0",
    "mypy>=1.13.0",
    "pre-commit>=4.0.0",
]

[project.scripts]
eab = "agent_builder.cli:main"
agent-builder = "agent_builder.cli:main"

[project.urls]
Homepage = "https://github.com/ai-solutions/easy-agent-builder"
Documentation = "https://docs.easy-agent-builder.ai"
Repository = "https://github.com/ai-solutions/easy-agent-builder"

[tool.hatch.build.targets.wheel]
packages = ["src/agent_builder"]

[tool.black]
line-length = 100
target-version = ['py311']

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "C4", "SIM"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
addopts = "-v --cov=agent_builder --cov-report=term-missing"
